2003-03-31  Bradford Hovinen  <bghovinen@math.uwaterloo.ca>
	    * tests/test-vector-domain.C (main): Add test for uint8

	    * linbox/field/modular.h (class FieldAXPY<uint8>): typedef uint8
	    Element

+++ linbox/linbox/field/modular.h
@@ -1039,7 +1039,7 @@ namespace LinBox
 	{
 	    public:
 
-		typedef uint16 Element;
+		typedef uint8 Element;
 		typedef Modular<uint8> Field;
 
 		FieldAXPY (const Field &F) : _F (F), i (F._k) { _y = 0; }
